% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsContainer.R
\name{acquisitionDirectory}
\alias{acquisitionDirectory}
\title{Create a table containg the classes of the acquisition.}
\usage{
acquisitionDirectory(files = NULL)
}
\arguments{
\item{files}{The path to the experiment.}
}
\value{
a table containing two columns, the absolute paths
of the files and the classes of the acquisition, as given by
subdirectories.
}
\description{
Find the classes oganization of a directory, and return a
table. This function is called by proFIAset, and is useful
to check the structure which will be stored in the proFIAset
object.
}
\examples{
if(require(plasFIA)){
    path<-system.file(package="plasFIA","mzML")
    tabClasses<-acquisitionDirectory(path)
    tabClasses
}
}

