% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsContainer.R
\name{analyzeAcquisitionFIA}
\alias{analyzeAcquisitionFIA}
\title{Wrapper function for the full FIA analysis workflow.}
\usage{
analyzeAcquisitionFIA(path, ppm, dmz = 0.001, fracGroup = 0.5,
  ppmGroup = NULL, dmzGroup = 5e-04, parallel = FALSE, bpparam = NULL,
  noiseEstimation = TRUE, SNT = NULL, maxo = FALSE,
  imputation = c("KNN_TN", "randomForest", "None"), ...)
}
\arguments{
\item{path}{The path to the directory of acquistion.}

\item{ppm}{The tolerance for deviations in m/z between scan in ppm \code{\link{findFIASignal}}}

\item{dmz}{The minimum tolerance for deviations in m/z between scan in Dalton \code{\link{findFIASignal}}}

\item{fracGroup}{The fraction of smaple form a class necessary to make a group.}

\item{ppmGroup}{A ppm tolerance to group signals between samples \code{\link{group.FIA}}.}

\item{dmzGroup}{The minimum tolerance in Dalton to group signals between samples \code{\link{group.FIA}}.}

\item{parallel}{A boolean indicating if parallelism is supposed to be used.}

\item{bpparam}{A BiocParallelParam object to be passed i BiocParallel is used.}

\item{noiseEstimation}{A boolean indicating in noise need to be estimated.}

\item{SNT}{A value giving the SNT thrshold, used only if \code{noiseEstimation} is FALSE.}

\item{maxo}{Should the maximum intensity be used over the peak area.}

\item{imputation}{The method to use for imputation 'randomForest' (\code{\link{impute.randomForest}})  or 'KNN_TN' (\code{\link{impute.KNN_TN}}). Put 'None' if  no imputation should be done.}

\item{...}{Supplementary arguments to be passed to the imputation method, see \code{\link{impute.FIA}} for detail.}
}
\value{
A filled proFIAset object ready for exportation.
}
\description{
Perform the 4 steps pro fia workflow including :
\itemize{
 \item noise estimation. Noise is estimated.
 \item bands filtering. Bands are filtered using the \code{\link{findFIASignal}} function.
 \item peak grouping. Signals from different acquisition are grouped using \code{\link{group.FIA}} function.
 \item missing values imputations. Missing values are imputated using \code{\link{impute.KNN_TN}} function.
}
Minimal options to launch the workflow are provided, neithertheless if finer option tuning are
necessary, launching the workflow function by function is strongly advised.
}
\examples{
if(require(plasFIA)){
    path<-system.file(package="plasFIA","mzML")
    
    #Defining parameters for Orbitrap fusion.
    ppm<-2
    dmz <- 0.0005
    ppm_group<-1
    dmz_group <- 0.0003
    paral<-FALSE
    frac_group<-0.2
    k<-2
    maxo<-FALSE
    vimputation <- "randomForest"
    \dontrun{plasSet<-analyzeAcquisitionFIA(path,ppm=ppm,dmz=dmz,imputation=vimputation,fracGroup=frac_group,ppmGroup=ppm_group,dmzGroup=dmz_group,k=k,maxo=maxo,parallel=paral)}
}
}
