% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPeaksFIA.R
\name{determiningInjectionZone}
\alias{determiningInjectionZone}
\title{First guess of the limit of the injection peak.}
\usage{
determiningInjectionZone(xraw, threshold = 0.05, graphical = FALSE,
  scanmin = NULL, scanmax = NULL)
}
\arguments{
\item{xraw}{An xcmsRaw object as returned by \code{\link[xcms]{xcmsRaw}}. It may also be
a TIC sequence as a list scan intensity.}

\item{threshold}{A relative increase born to detect the limit of the injection
peak.}

\item{graphical}{should the resulting limit be plotted.}

\item{scanmin}{The first scan to consider.}

\item{scanmax}{The last scan to consider.}
}
\value{
A triplet composed of c(left limit,right limit, maximum) of the
estimated injection peak.
}
\description{
Determine a first approximation of the injection peak based
on the angle of the injection peak and changing of variation.
This peak is not used directly by findFIAsignal, but will be used to initialize the regression
giving the injection peak. Shloud be used carefuly.
}
\examples{
if(require(plasFIA)){
  #Getting the path of a file.
  path_raw <- list.files(system.file(package="plasFIA","mzML"),full.names=TRUE)[2]

  #Opening the file with xcms
  xraw <- xcmsRaw(path_raw)

  #Getting a first approximation of injection peak;
  sp <- determiningInjectionZone(xraw)
}
}
