% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPeaksFIA.R
\name{determiningSizePeak.Geom}
\alias{determiningSizePeak}
\alias{determiningSizePeak.Geom}
\title{Determine the limits of the injection peak in a FIA acquisition.}
\usage{
determiningSizePeak.Geom(xraw, scanmin = 1, scanmax = length(xraw@scantime),
  freq = 0.15, graphical = FALSE, smooth = TRUE, extended = FALSE,
  percentSol = NULL)
}
\arguments{
\item{xraw}{An xcmsRaw object as returned by \code{\link[xcms]{xcmsRaw}}.}

\item{scanmin}{The minimum scan to consider for peak detection.}

\item{scanmax}{the maximum scan to consider.
injection peak detection.}

\item{freq}{The degrees of smoothing used on the TIC, corresponding to
the cutting frequency of the blackman windowed sync filter.}

\item{graphical}{should the resulting peak be plotted.}

\item{smooth}{Should the TIC be smoothed, recommended.}

\item{extended}{In case of very long tailing, shloud the research be extended.}

\item{percentSol}{If extended is TRUE, the limiting level of solvent for}
}
\value{
A triplet composed of c(left limit,right limit, maximum) of the
estimated injection peak.
}
\description{
Determine a first approximation of the injection peak using the
Douglas-Peuker Algorithm provided in the \code{rgeos} package.
The object provided must be an xcmsRaw object.
}
\examples{
if(require(plasFIA)){
  #Getting the path of a file.
  path_raw <- list.files(system.file(package="plasFIA","mzML"),full.names=TRUE)[2]

  #Opening the file with xcms
  xraw <- xcmsRaw(path_raw)

  #Getting a first approximation of injection peak;
  sp <- determiningSizePeak.Geom(xraw)
}
}

