% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noiseEstimator.R
\name{estimateNoiseMS}
\alias{estimateNoiseMS}
\title{Estimate the noise of a mass spectrometer using multiple MS acquisition.}
\usage{
estimateNoiseMS(list_files, ppm, nBin = 500, minInt = 500, maxInt = 10^8,
  parallel, BPPARAM = NULL)
}
\arguments{
\item{list_files}{A list of files in which the noise should be estimated.}

\item{ppm}{The authorized deviation between scans in ppm, this parameter
will also be used to fuse the bands if there are close enough.}

\item{nBin}{The number of intensity bins to be used.}

\item{minInt}{The  minimum intensity expected in all the files.}

\item{maxInt}{The  maximum intensity expected in all the files.}

\item{parallel}{Shall parrallelism be used.}

\item{BPPARAM}{A BiocParallelParam object to be used for parallelism
if parallel is TRUE.}
}
\value{
A  noise estimator object.
}
\description{
Determine the variances of the noise in function of the intensity
from multiples FIA acquisitions, using the method from Wentzell and Tarazuk(2014)
\emph{Characterization of heteroscedastic measurement noise in the absence of replicates}
}
\examples{
##Listing the files in plasFIA
if(require(plasFIA)){
  list_mzml <- list.files(system.file(package="plasFIA","mzML"),full.names=TRUE)

  ##For speed purpose
  list_mzml <- list_mzml[1:2]
  es <- estimateNoiseMS(list_mzml,2,parallel=FALSE)
}
}

