% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsContainer.R
\docType{methods}
\name{exportDataMatrix,proFIAset-method}
\alias{exportDataMatrix}
\alias{exportDataMatrix,proFIAset-method}
\title{Export data matrix.}
\usage{
\S4method{exportDataMatrix}{proFIAset}(object, filename = NULL)
}
\arguments{
\item{object}{A proFIAset object.}

\item{filename}{If not NULL the result will be written
in filename as a tabular separated values file.}
}
\value{
A matrix with dimension samples x variables.
}
\description{
Export the data matrix from a \code{\link{proFIAset}} object, to
be used for statistical analysis.
}
\examples{
if(require(plasFIA)){
    data(plasSet)
    dm<-exportDataMatrix(plasSet)
    head(dm)
}
}

