% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsContainer.R
\docType{methods}
\name{exportExpressionSet,proFIAset-method}
\alias{exportExpressionSet}
\alias{exportExpressionSet,proFIAset-method}
\title{Export proFIAset to ExpressionSet}
\usage{
\S4method{exportExpressionSet}{proFIAset}(object, colgroup = c("mzMed",
  "scanMin", "scanMax", "nPeaks", "corMean", "SigSolMean", "shifted"))
}
\arguments{
\item{object}{A proFIAset object.}

\item{colgroup}{Labels corresponding to the column names
of the group table.}
}
\value{
An \code{ExpressionSet} object from the \code{Biobase} package
}
\description{
Eport the data from a proFIAset object as an \code{ExpressionSet} object
from the \code{Biobase} package package.
}
\examples{
if(require("plasFIA")&require("Biobase")){
    data(plasSet)
    eset<-exportExpressionSet(plasSet)
    eset
}
}

