% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsContainer.R
\docType{methods}
\name{exportPeakTable,proFIAset-method}
\alias{exportPeakTable,proFIAset-method}
\alias{exportPeakTable}
\title{Export proFIAset as a peak table.}
\usage{
\S4method{exportPeakTable}{proFIAset}(object, colgroup = c("mzMed",
  "corSampPeakMean", "meanSolvent", "signalOverSolventRatioMean", "timeShifted",
  "signalOverSolventPvalueMean"), mval = c("NA", "zero"), filename = NULL)
}
\arguments{
\item{object}{A proFIAset object.}

\item{colgroup}{Labels corresponding to the column names
of the group table which will be added to the peak table.}

\item{mval}{How will missing values be treated, in default they
will be set to NA, or you can keep 0.}

\item{filename}{The name of the file for the peak table to be exported.}
}
\value{
A dataframe containg the datasets.
}
\description{
Export the data from a proFIAset object as
a peak table which containes the values
of measured for each variables for each samples
and supplementary information.
}
\examples{
if(require("plasFIA")){
    data(plasSet)

    #Creating the peak table
    ptable<-exportPeakTable(plasSet)
    head(ptable)

    #Directly in a file
    \dontrun{ptable<-exportPeakTable(plasSet,filename="peak_table.tsv")}
}
}
