% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsContainer.R
\docType{methods}
\name{exportSampleMetadata,proFIAset-method}
\alias{exportSampleMetadata}
\alias{exportSampleMetadata,proFIAset-method}
\title{Export samples metadata.}
\usage{
\S4method{exportSampleMetadata}{proFIAset}(object, filename = NULL)
}
\arguments{
\item{object}{A proFIAset object.}

\item{filename}{If not NULL the result will be written
in filename}
}
\value{
A dataframe with the following columns :
\itemize{
    \item sampleID an ID similar to the one of the peak table.
    \item class the group of the sample.
}
}
\description{
Export the samples metadata of an experiment, to
be used for statistical analysis.
}
\examples{
if(require(plasFIA)){
   data(plasSet)
   tsample<-exportSampleMetadata(plasSet)
   head(tsample)
}
}

