% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsContainer.R
\docType{methods}
\name{fillPeaks.WKNN,proFIAset-method}
\alias{fillPeaks.WKNN}
\alias{fillPeaks.WKNN,proFIAset-method}
\title{Fill missing values in the peak table.}
\usage{
\S4method{fillPeaks.WKNN}{proFIAset}(object, k = 5)
}
\arguments{
\item{object}{A proFIAset object.}

\item{k}{The number of neighbors considered.}
}
\value{
A proFIAset object with the missing values imputated.
}
\description{
Impute the missing values in an FIA experiment using a Weighted
K-Nearest Neighbours.
}
\examples{
if(require(plasFIA)){
    data(plasSet)
    
    ###Reinitializing the data matrix
    plasSet<-makeDataMatrix(plasSet,maxo=FALSE)
    plasSet<-fillPeaks.WKNN(plasSet,2)
}
}

