% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPeaksFIA.R
\name{findFIASignal}
\alias{findFIASignal}
\alias{findPeaks}
\title{Detect peaks in an FIA acquisition.}
\usage{
findFIASignal(xraw, ppm, es = NULL, solvar = c("throw", "keep"),
  solint = c("poly", "substract", "keep"), dmz = 5e-04, graphical = FALSE,
  SNT = NULL, f = c("regression", "TIC"), pvalthresh = NULL, ...)
}
\arguments{
\item{xraw}{An xcmsRaw object as returned by \code{\link[xcms]{xcmsRaw}}.}

\item{ppm}{The authorized deviation between scans in ppm, this parameter
will also be used to fuse the bands if there are close enough.}

\item{es}{A noise estimation object as returned by \link{estimateNoiseListFiles}, or NULL
if the parameter noise if only an threshold is supposed to be used.}

\item{solvar}{Should the signal corresponding to solvent be kept ?
Only their maximum intensiyt will be calculated.}

\item{solint}{How are the intensity of signal with bot solvent and sample
be treated in the injection zone region, the area of the rectangle
with peak-width and solvent intensity is considered.
\itemize{
    \item poly. Half of this area is kept.
    \item substract. The area is removed substracted.
    \item remove. The area is conserved in the final value.
}}

\item{dmz}{The minimum absolute value of the deviation between scans, to take into account
the higher diviations at low masses.}

\item{graphical}{A boolean indicating if the detected area shall be plotted.}

\item{SNT}{NULL by default a relative intensity of signal/intensity of solvent threshold,
used only if es is equal to NULL.}

\item{f}{method to design the filter, "TIC" means that the peak of the TIC is
used as a filter. "regression" means that the signal is regressed form the most
intense band as an Exponential modified gaussian.}

\item{pvalthresh}{The threshold used in p-value to discard signal, only used if
a noise model is furnished.}

\item{...}{more arguments to be passed to the \link{determiningSizePeak} function.}
}
\value{
A numeric matrix with the following column
\itemize{
    \item mzmin the minimum value of the mass traces in the m/z dimension.
    \item mzmax the maximum value of the mass traces in the m/z dimension.
    \item scanMin the first scan on which the signal is detected.
    \item scanMax the last scan on which the signal is detected.
    \item areaIntensity the integrated area of the signal.
    \item maxIntensity the maximum intensity of the signal.
    \item solventIntensity the intensity of the solvent, 0 means that no significant
    solvent was detected.
    \item corPeak An idicator of matrix effect, if it's close to 1, the compound
    does not suffer from heavy matrix effect, if it is inferior to 0.5, the compound
    suffer from heavy matrix effect.
    \item Sig_Sol The ratio of the signal max intensity on the oslvent max intensity.
    }
}
\description{
Detect the peak corresponding to compounds present in the sample
in a Flow Injection Analysis (FIA) acquisition. The item furnished
must be an xcmsRaw object.
}
\examples{
if(require(plasFIA)){
  #Getting the path of a file.
  path_raw<-list.files(system.file(package="plasFIA","mzML"),full.names=TRUE)[2]

  #Opening the file with xcms
  xraw<-xcmsRaw(path_raw)

  ppm<-2

  #getting the filtered signals without noise model which is not recommended.
  tsignal<-findFIASignal(xraw,ppm=ppm,SNT=3)

  #Getting the noise model un the plasSet object.
  data(plasSet)
  es<-attr(plasSet,"noiseEstimation")

  #Getting the signal with a noise model. 
  tsignal<-findFIASignal(xraw,ppm=2,es=es,pvalthresh=0.005)
  head(tsignal)
}
}

