% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsContainer.R
\docType{methods}
\name{findMzGroup,proFIAset-method}
\alias{findMzGroup}
\alias{findMzGroup,proFIAset-method}
\title{find a group in a FIA experiment.}
\usage{
\S4method{findMzGroup}{proFIAset}(object, mz, tol = 0)
}
\arguments{
\item{object}{A proFIAset object.}

\item{mz}{A numeric vector of masses to be looked for.}

\item{tol}{The tolerance in ppm.}
}
\value{
A vector of integer of the same length than mz giving the row
of the found group in the object group slot, or NA if the group is
not found.
}
\description{
Find a group corresponding to the given mass in a proFIAset
object.
}
\examples{
if(require("plasFIA")){
#proFIAset object is loaded
data(plasSet)

#The table of spiked molecule is loaded
data(plasMols)

#Mass to search and toolerance are defined
mass<-plasMols[22,"mass_M+H"]
tolppm<-1

plasSet=makeDataMatrix(plasSet)

index=findMzGroup(plasSet,mass,tol=1)
plasSet
}
}
\seealso{
You can visualize the group using \code{\link{plotFlowgrams}} function.
}

