% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPeaksFIA.R
\name{getInjectionPeak}
\alias{getInjectionPeak}
\title{Fit an injection peak to an FIA acquisition.}
\usage{
getInjectionPeak(xraw, bandlist, sec = 2, iquant = 0.95, gpeak = NULL,
  selIndex = NULL, graphical = FALSE)
}
\arguments{
\item{xraw}{An xcmsRaw object as returned by \code{\link[xcms]{xcmsRaw}}.}

\item{bandlist}{A list of bands which can be used. Shall stay NULL in general use.
bands will be determined automatically.}

\item{sec}{A tolerance in sec to group the signals.}

\item{iquant}{The maximum intensity intensity threshold under which the peaks
would not be used for peak determination.}

\item{gpeak}{An approximation of the injection peak, if NULL
determining sizepeak.Geom will be used.}

\item{selIndex}{A seleciton of index to make the regression. We recommend to let it to NULL.}

\item{graphical}{shald the individually fitted components be plotted.}
}
\value{
A vector contaning the injection peak
}
\description{
Determine an injection peak as an exponential modified gaussian
function and a second order exponential corresponding to matrix
effect to the most intense signals in an acquisition.
}
\examples{
if(require(plasFIA)){
  #Getting the path of a file.
  path_raw <- list.files(system.file(package="plasFIA","mzML"),full.names=TRUE)[2]

  #Opening the file with xcms
  xraw <- xcmsRaw(path_raw)

  #Getting the injection scan
  gp <- determiningInjectionZone(xraw)

  #performing band detection.
  tbands <- findBandsFIA(xraw,ppm = 2,sizeMin = gp[3]-gp[1],beginning=gp[1],end=gp[2])

  #Getting the injection peak
  vpeak <- getInjectionPeak(xraw,bandlist=tbands,gpeak=gp)
  plot(vpeak,type="l")
}
}
