% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsContainer.R
\docType{methods}
\name{group.FIA,proFIAset-method}
\alias{group.FIA,proFIAset-method}
\alias{group.FIA}
\title{Group the peaks of an FIA acquisition.}
\usage{
\S4method{group.FIA}{proFIAset}(object, ppmGroup, solvar = FALSE, sleep = 0,
  dmzGroup = 5e-04, fracGroup = 0.5)
}
\arguments{
\item{object}{A proFIAset object.}

\item{ppmGroup}{A ppm parameter giving the size of the windows
considered, we recommend to use 0.5*ppm where ppm is the pp parameter
of band detection in the \code{\link{proFIAset}} function.}

\item{solvar}{Shall the group corresponding to solvent signal be kept. This in only useful if solvent
signal have been conserved in the \code{\link{findFIASignal}} function.}

\item{sleep}{If not 0 densities are plotted every \code{sleep} ms.}

\item{dmzGroup}{A minimum mz deviation value which can be considered over the deviation in ppm
if it is higher. This account for low mass deviation.}

\item{fracGroup}{The minimum fraction of samples of a class required to make
a group.}
}
\value{
A proFIAset object with the \code{group} slot filled. See \code{\link{proFIAset-class}}.
}
\description{
Group the peaks in a FIA experiemnts by clustering under
an estimated density based on the accuracy in ppm
of the mass spectrometer.
}
\examples{
if(require("plasFIA")){
#proFIAset object is loaded
data(plasSet)

#Parameters are defined.
ppm_group <- 1
dmz_group <- 0.0005
frac_group<-0.2

plasSet<-group.FIA(plasSet,ppmGroup=ppm_group,fracGroup=frac_group,dmzGroup=dmz_group)
plasSet
}
}
