% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsContainer.R
\docType{methods}
\name{impute.FIA,proFIAset-method}
\alias{impute.FIA,proFIAset-method}
\alias{impute.FIA}
\title{Fill missing values using the provided method.}
\usage{
\S4method{impute.FIA}{proFIAset}(object, method = c("KNN_TN", "randomForest"),
  ...)
}
\arguments{
\item{object}{A proFIAset object.}

\item{method}{The method to be used for missing value imputation.}

\item{...}{Arguments furnished to the imputation method. No argument is required for \code{\link{impute.randomForest}} and the number of neighbours should be furnished using the \code{\link{impute.KNN_TN}} function.}
}
\description{
Impute the missing values in an FIA experiment using a Weighted
K-Nearest Neighbours on Truncated Distribution implemented in \code{\link{impute.KNN_TN}} or by random forest using the \code{\link{impute.randomForest}} function.
}
\examples{
if(require(plasFIA)){
    data(plasSet)

    ###Reinitializing the data matrix an using KNN
    plasSet<-makeDataMatrix(plasSet,maxo=FALSE)
    plasSet<-impute.FIA(plasSet,method="KNN_TN",k=2)
    
    ###Reinitializing the data matrix and using randomForest
    plasSet<-makeDataMatrix(plasSet,maxo=FALSE)
    plasSet<-impute.FIA(plasSet,method="randomForest")
}
}
