% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsContainer.R
\docType{methods}
\name{impute.randomForest,proFIAset-method}
\alias{impute.randomForest,proFIAset-method}
\alias{impute.randomForest}
\title{Fill missing values in the peak table using random forest.}
\usage{
\S4method{impute.randomForest}{proFIAset}(object, parallel = FALSE, ...)
}
\arguments{
\item{object}{A proFIAset object.}

\item{parallel}{Shall parallelism be used.}

\item{...}{supplementary arguements to be passed to missForest values.}
}
\value{
A proFIAset object with the missing values imputated.
}
\description{
Impute the missing values in an FIA experiment using a random forest implemented
in the missForest package.
}
\examples{
if(require(plasFIA)){
    data(plasSet)
    ###Reinitializing the data matrix
    plasSet<-makeDataMatrix(plasSet,maxo=FALSE)
    plasSet<-impute.randomForest(plasSet)
}
}
\references{
Stekhoven, D.J. and Buehlmann, P. (2012), 'MissForest - nonparametric missing value imputation for mixed-type data',
 Bioinformatics, 28(1) 2012, 112-118, doi: 10.1093/bioinformatics/btr597
}
