% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsContainer.R
\docType{methods}
\name{makeDataMatrix,proFIAset-method}
\alias{makeDataMatrix,proFIAset-method}
\alias{makeDataMatrix}
\title{Construct the data matrix of a proFIAset object.}
\usage{
\S4method{makeDataMatrix}{proFIAset}(object, maxo = FALSE)
}
\arguments{
\item{object}{A \code{\link{proFIAset}} object.}

\item{maxo}{Shall the intensity used to the area or the maximum
intensity}
}
\value{
A proFIAset object with the \code{dataMatrix} slot filled.
}
\description{
Construct the data matrix of a proFIA set object, using the selected measure for the intensity,
area or maximum intensity. The choice of area or maximum intensity depends of your acquisition,
and your preferences.
}
\examples{
if(require("plasFIA")){
#proFIAset object is loaded
data(plasSet)

plasSet<-makeDataMatrix(plasSet)
plasSet
}
}
\seealso{
To obtain this data matrix see \code{\link{proFIAset}}.
}
