% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noiseEstimator.R
\docType{class}
\name{noiseEstimation-class}
\alias{noiseEstimation-class}
\title{An S4 class to represent heteroscedasctic noise of MS.}
\description{
An S4 class to represent heteroscedasctic noise of MS.
}
\section{Slots}{

\describe{
\item{\code{bins}}{The limit of the bins on which the noise have been estimated.}

\item{\code{varmean}}{The estimate of the mean of the variance in each bin.}

\item{\code{size}}{The size of each bins in number of elements.}

\item{\code{estimation}}{The estimation function, if a model have been fitted.}

\item{\code{filelist}}{The list of the fitted file.}

\item{\code{estimated}}{A boolean indicating if a model have been fitted.}

\item{\code{frac}}{A numeric giving the fraction of the point ot be conserved when estimating the noise model.}

\item{\code{intlim}}{The maximum and minmum intensity on which the estimationis done.}

\item{\code{reglim}}{The limits of the regression of the model on these data.
detected for each experiment.}
}}

