% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsContainer.R
\docType{methods}
\name{plot,proFIAset,ANY-method}
\alias{plot,proFIAset,ANY-method}
\alias{plot.FIA}
\title{Plot a summary of an FIA experiment.}
\usage{
\S4method{plot}{proFIAset,ANY}(x, y = NULL, type = c("sample", "class"))
}
\arguments{
\item{x}{A proFIAset object.}

\item{y}{Unused at the moment.}

\item{type}{Shall the plotting be done by sample or by class for the barplot ?}
}
\value{
Nothing
}
\description{
Plot a summary of an FIA acquisition. This summary aims to provides an overview of the
FIA acquisition and the processinf of FIA acquisition. It includes the following graphs :
\itemize{
    \item Barplot A barplot giving the number of peak detected in each sample. The number 
    of shifted peak is indicated, which can indicate wrong FIA-acquisition, as well as the
    number of peak badly with a low correlation with the injection peak, which can indicate
    a strong matrix effect.
    \item Injection_Peaks A representation of all the injection peaks of the acquisition, a
    greatly different peak in th einjection may indicate an issue with the injection, or a problem
    of regression. If the peaks seems all different, try to use \link{proFIAset} with the f
    paramter on TIC, to avoid regression.
    \item Density A density plot of the m/z of all the features found. A missing interval at the end
    of the mz range may indicate a too low ppm parameter, while a missing interval at the beginning
    of the mz range may indicate a too low dmz parameter.
    \item PCA A PCA plot to obtain a quick diagnostic of the data. The PCA plot is supposed to be different
    before and after imputation.
}
}
\examples{
if(require("plasFIA")){
   data(plasSet)
   
   
   ####Diagnostic plot after imputation
   plot(plasSet)
   
   ####The same plot by classes.
   plot(plasSet,type="class")
   
   ####Diagnostic plot before imputation
   plasSet <- makeDataMatrix(plasSet)
  plot(plasSet)

}
}

