% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsContainer.R
\docType{methods}
\name{plotInjectionPeaks,proFIAset-method}
\alias{plotInjectionPeaks}
\alias{plotInjectionPeaks,proFIAset-method}
\title{Plot the injection peaks of a proFIAset object.}
\usage{
\S4method{plotInjectionPeaks}{proFIAset}(object, subsample = NULL, ...)
}
\arguments{
\item{object}{A proFIAset object.}

\item{subsample}{The subset of sample on which the sample may be plotted.
If it is numeric it will be viewed as sample row in the classes table,
if it is a character it will be viewed as a factor.}

\item{...}{SUpplementary arguments which will be passed to the \link{lines}
function.}
}
\value{
No value returned.
}
\description{
Plot the injection peaks evaluated on each raw files
of a proFIAset object one plot. If a peak is really
different from the other it can indicater an error in
the injection process.
}
\examples{
if(require(plasFIA)){
    data(plasSet)
    plotInjectionPeaks(plasSet)
}
}

