% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noiseEstimator.R
\name{plotNoise}
\alias{plotNoise}
\title{Plot the estimated noise from a proFIAset object.}
\usage{
plotNoise(object, xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{object}{A noise estimation object or a \code{\link{proFIAset-class}} object.}

\item{xlim}{The xlim paramter to be passed to plot.}

\item{ylim}{The ylim paramter to be passed to plot.}

\item{...}{SUpplementary arguments to be passed to plot}

\item{plotNoise}{plotNoise,proFIAset-method}
}
\value{
The plotted value as an x y list.
}
\description{
Plot an intensity vs variances plot for the noise estimated from a
MS acquisition. If a model is fitted to the data it will be plotted.
Only the interval used for the regression and on which the estimation will be used is shown.
}
\examples{
if(require(plasFIA)){
  data(plasSet)
  plotNoise(plasSet)
}
}
