% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsContainer.R
\docType{methods}
\name{plotRaw,proFIAset-method}
\alias{plotRaw,proFIAset-method}
\alias{plotRaw}
\title{Plotting of raw data}
\usage{
\S4method{plotRaw}{proFIAset}(object, type = c("raw", "peaks"),
  sample = NULL, ...)
}
\arguments{
\item{object}{A proFIAset object.}

\item{type}{\code{"raw"} indicate that raw data needs to be plotted
and \code{"peak"} indicate that only the filtered signals will be plotted.}

\item{sample}{The numpber of the sample in the object classes table
to be plotted.}

\item{...}{xlim,ylim and size to be passed to plot functions.}
}
\value{
No value is returned.
}
\description{
Plot the raw data from a proFIAset object,the
the type of plot determines if the full raw data
needs to be plotted or only the data conressponding to
the detected peaks needs to be plottes. The path in the classes
table of the proFIAset object needs to be correct.
}
\examples{
if(require("plasFIA")){
    data(plasSet)

    #Visualising the raw data
    plotRaw(plasSet,type="raw",ylim=c(215.9,216.2),sample=4)

    #Plotting the filtered signals only.
    plotRaw(plasSet,type="peaks",ylim=c(215.9,216.2),sample=4)
}
}
