% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsContainer.R
\docType{methods}
\name{plotSamplePeaks,proFIAset-method}
\alias{plotSamplePeaks,proFIAset-method}
\alias{plotSamplePeaks}
\title{Plot the sample peak of a proFIAset object.}
\usage{
\S4method{plotSamplePeaks}{proFIAset}(object, subsample = NULL,
  diagPlotL = FALSE, ...)
}
\arguments{
\item{object}{A proFIAset object.}

\item{subsample}{The subset of sample on which the sample may be plotted.
If it is numeric it will be viewed as sample row in the classes table,
if it is a character it will be viewed as a factor.}

\item{diagPlotL}{Boolean used by the usmmary plot function. Keep to FALSE in the general case.}

\item{...}{SUpplementary arguments which will be passed to the \link{lines}
function.}
}
\value{
No value returned.
}
\description{
Plot the sample peaks determined by regression for each sample.
If you want to check the various regression performed by
\emph{proFIA} in an individual sample we recommend you to use the
\code{\link{getInjectionPeak}}. A sample peak highly different form the 
other may indicate a problem of the regression process, or an injection issue
in the acquisition.
}
\examples{
if(require(plasFIA)){
    data(plasSet)
    plotSamplePeaks(plasSet)
}
}
