% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classContainer.R, R/methodsContainer.R
\docType{class}
\name{proFIAset-class}
\alias{dataMatrix}
\alias{dataMatrix,proFIAset-method}
\alias{groupMatrix}
\alias{groupMatrix,proFIAset-method}
\alias{injectionPeaks}
\alias{injectionPeaks,proFIAset-method}
\alias{peaks}
\alias{peaks,proFIAset-method}
\alias{phenoClasses}
\alias{phenoClasses,proFIAset-method}
\alias{proFIAset-class}
\title{An S4 class to represent an FIA experiments.}
\usage{
\S4method{phenoClasses}{proFIAset}(object)

\S4method{dataMatrix}{proFIAset}(object)

\S4method{groupMatrix}{proFIAset}(object)

\S4method{peaks}{proFIAset}(object)

\S4method{injectionPeaks}{proFIAset}(object)
}
\arguments{
\item{object}{A proFIAset object.}
}
\description{
The S4 class also includes all the informations about processing, and the detected signals
are stored.
}
\section{Methods (by generic)}{
\itemize{
\item \code{phenoClasses}: Extract the classes and the paths of the samples.

\item \code{dataMatrix}: Extract the dataMatrix containg variables as rows and samples as columns

\item \code{groupMatrix}: Extract the matrix of group, see \code{\link{exportPeakTable}} for better output.

\item \code{peaks}: Extract all the signals detected in individual samples.

\item \code{injectionPeaks}: Extract all the regressed injection peaks.
}}
\section{Slots}{

\describe{
\item{\code{peaks}}{A matrix containg all the peaks which have been detected in each
individual files.
\itemize{
    \item mzmin the minimum value of the mass traces in the m/z dimension.
    \item mzmax the maximum value of the mass traces in the m/z dimension.
    \item scanMin the first scan on which the signal is detected.
    \item scanMax the last scan on which the signal is detected.
    \item areaIntensity the integrated area of the signal.
    \item maxIntensity the maximum intensity of the signal.
    \item solventIntensity the intensity of the solvent, 0 means that no significant
    solvent was detected.
    \item corPeak An idicator of matrix effect, if it's close to 1, the compound
    does not suffer from heavy matrix effect, if it is inferior to 0.5, the compound
    suffer from heavy matrix effect.
    \item Sig_Sol The ratio of the signal max intensity on the oslvent max intensity.
}}

\item{\code{group}}{A matrix containing the information on the groups done between all the
acquisitions.
\itemize{
    \item mzMed the median value of group in the m/z dimension.
    \item mzMin the minimum value of the group in the m/z dimension.
    \item mzMax the maximum value of the group in the m/z dimension.
    \item scanMin the first scan on which the signal is detected.
    \item scanMax the last scan on which the signal is detected.
    \item nPeaks The number of peaks grouped in a group.
    \item meanSolvent The mean of solvent in the acquisition.
    \item pvalueMean The mean p-value of the group.
    \item corMean The mean of the matrix effect indicator.
    \item SigSolMean The mean of ratio of the signal max
    intensity on the oslvent max intensity.
    \item corSd The standard deviation of the matrix effect indicator.
}}

\item{\code{groupidx}}{The row of the peaks corresponding to each group
in \code{peaks}.}

\item{\code{step}}{The step of processing of the experiment.}

\item{\code{path}}{The path of the experiment.}

\item{\code{classes}}{A table with two columns, "rname" the absolute path
of a file, and "group" the class to which the file belong.}

\item{\code{dataMatrix}}{A matrix variables x experiments suitable for
statistical analysis.}

\item{\code{noiseEstimation}}{A model of noise as estimated by
\link{estimateNoiseListFiles}}

\item{\code{injectionPeaks}}{A list of the injection peak which have been
detected for each experiment.}

\item{\code{injectionScan}}{A numeric vector giving the scan of injection of
sample.}
}}

