% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsContainer.R
\name{analyzeAcquisitionFIA}
\alias{analyzeAcquisitionFIA}
\title{Wrapper function for the full FIA analysis workflow.}
\usage{
analyzeAcquisitionFIA(path, ppm, fracGroup = 0.5, ppmgroup = NULL,
  parallel = FALSE, bpparam = NULL, noiseEstimation = TRUE, SNT = NULL,
  maxo = FALSE, k = 5)
}
\arguments{
\item{path}{The path to the directory of acquistion.}

\item{ppm}{The tolerance for deviations in m/z between scan in ppm \code{\link{findFIASignal}}}

\item{fracGroup}{The fraction of smaple form a class necessary to make a group.}

\item{ppmgroup}{A ppm tolerance to group signals between samples \code{\link{group.FIA}}.}

\item{parallel}{A boolean indicating if parallelism is supposed to be used.}

\item{bpparam}{A BoicParallelParam object to be passed i BiocParallel is used.}

\item{noiseEstimation}{A boolean indicating in noise need to be estimated.}

\item{SNT}{A value giving the SNT thrshold, used only if \code{noiseEstimation} is FALSE.}

\item{maxo}{Should the maximum intensity be used over the peak area.}

\item{k}{The number of neighbors for \code{\link{fillPeaks.WKNN}}.}
}
\value{
A filled proFIAset object ready for exportation.
}
\description{
Perform the 4 steps pro fia workflow including :
\itemize{
 \item noise estimation. Noise is estimated.
 \item bands filtering. Bands are filtered using the \code{\link{findFIASignal}} function.
 \item peak grouping. Signals from different acquisition are grouped using \code{\link{group.FIA}} function.
 \item missing values imputations. Missing values are imputated using \code{\link{fillPeaks.WKNN}} function.
}
Minimal options to launch the workflow are provided, neithertheless if finer option tuning are
necessary, launching the workflow function by function is strongly advised.
}
\examples{
if(require(plasFIA)){
    path<-system.file(package="plasFIA","mzML")
    
    #Defining parameters for Orbitrap fusion.
    ppm<-2
    ppmgroup<-1
    paral<-FALSE
    fracGroup<-0.2
    k<-2
    maxo<-FALSE
    
    \dontrun{plasSet<-analyzeAcquisitionFIA(path,ppm=ppm,fracGroup=fracGroup,ppmgroup=ppmgroup,k=k,parallel=paral)}

}
}

