% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsContainer.R
\docType{methods}
\name{group.FIA,proFIAset-method}
\alias{group.FIA}
\alias{group.FIA,proFIAset-method}
\title{Group the peaks of an FIA acquisition.}
\usage{
\S4method{group.FIA}{proFIAset}(object, ppm, solvar = FALSE, nPoints = 1024,
  sleep = 0, fracGroup = 0.5)
}
\arguments{
\item{object}{A proFIAset object.}

\item{ppm}{The estimated accuraccy of the mass spectrometer
given in ppm, this parameter is supposed to be smaller
than the ppm parameter than in the \code{\link{proFIAset}} function}

\item{solvar}{Shall the group corresponding to solvent signal be kept.
Default is no.}

\item{nPoints}{the number of points used on the density, this parameter
only needs to be changed if the number of group found seems small.}

\item{sleep}{If not 0 densities are plotted every \code{sleep} ms.}

\item{fracGroup}{The minimum fraction of samples of a class required to make
a group.}
}
\value{
A proFIAset object with the \code{group} slot filled. See \code{\link{proFIAset-class}}.
}
\description{
Group the peaks in a FIA experiemnts by clustering under
an estimated density based on the accuracy in ppm
of the mass spectrometer.
}
\examples{
if(require("plasFIA")){
#proFIAset object is loaded
data(plasSet)

#Parameters are defined.
ppmgroup<-1
fracGroup<-0.2

plasSet<-group.FIA(plasSet,ppm=ppmgroup,fracGroup=fracGroup)
plasSet
}
}

