% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsContainer.R
\docType{methods}
\name{peaksGroup,proFIAset-method}
\alias{peaksGroup}
\alias{peaksGroup,proFIAset-method}
\title{Return the peaks corresponding to a group.}
\usage{
\S4method{peaksGroup}{proFIAset}(object, index = NULL)
}
\arguments{
\item{object}{A proFIAset object.}

\item{index}{A numeric vector r giving the group to be returned. NA are ignored.}
}
\value{
The peaks in the given group, see \code{\link{proFIAset-class}}.
}
\description{
Return the peaks corresponding ot a gorup given by his index.
}
\examples{
if(require(plasFIA)){
    data(plasSet)
    data(plasMols)
    
    #finding the molecules of plasMols
    vmatch<-findMzGroup(plasSet,mz=plasMols[,"mass_M+H"])

    mol_peaks<-peaksGroup(plasSet,index=vmatch)
    head(mol_peaks)
}
}

