% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsContainer.R
\name{proFIAset}
\alias{proFIAset}
\title{Process FIA experiment.}
\usage{
proFIAset(path, ppm, parallel = TRUE, BPPARAM = NULL,
  noiseEstimation = TRUE, graphical = FALSE, ...)
}
\arguments{
\item{path}{The path of the files to be processed.}

\item{ppm}{The tolerance of the algorithms in deviation between scans.}

\item{parallel}{Shall parallelism using \pkg{BiocParallel} be used.}

\item{BPPARAM}{A BiocParallelParam object to be used for parallelism
if parallel is TRUE.}

\item{noiseEstimation}{Shall noise be estimated (recommended)}

\item{graphical}{Shall the plot of the regressed injection peak be shown.}

\item{...}{Supplementary arguments to be passed to findFIAsignal, see
\code{\link{findFIASignal}}.}
}
\value{
A proFIAset object.
}
\description{
Processes an experiment ordered as a tree of files, and return a
proFIAset object. TO check the funirshed structure you may use the 
\code{\link{acquisitionDirectory}} function.
}
\examples{
if(require("plasFIA")){

pathplas<-system.file(package="plasFIA","mzML")

#Parameters are defined.
ppm<-2

plasSet<-proFIAset(pathplas,ppm=ppm,parallel=FALSE)
plasSet

###An example using parallelism with a snow cluster using BiocParallel package.
\dontrun{plasSet<-proFIAset(pathplas,ppm=ppm,parallel=TRUE,BPPARAM=bpparam("SnowParam"))}
}
}
\seealso{
To obtain more detail about the output see \code{\link{proFIAset-class}}.
}
